# app/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IT system to operate a clinic network",
        "version": "0.6.2",
        "title": "Clinic network system",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "michal.balicki@onet.pl"
        }
    },
    "host": "https://komponentowe-backend.herokuapp.com/",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/patient/:login": {
            "get": {
                "tags": [
                    "patient"
                ],
                "summary": "patient found by login",
                "description": "Requires apikey",
                "operationId": "patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/loginString"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "patient"
                ],
                "summary": "patient deleted by login",
                "description": "Requires apikey",
                "operationId": "patientDelete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/loginString"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/patient/:pesel": {
            "get": {
                "tags": [
                    "patient"
                ],
                "summary": "patient found by PESEL",
                "description": "Requires apikey",
                "operationId": "patientPesel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "pesel",
                        "description": "Pesel of patient to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/patient/signup": {
            "post": {
                "tags": [
                    "patient"
                ],
                "summary": "Patient register",
                "description": "Requires apikey",
                "operationId": "patientSignup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/patientSignup"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/patient/login": {
            "post": {
                "tags": [
                    "patient"
                ],
                "summary": "Patient login",
                "description": "Requires apikey",
                "operationId": "patientLogin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/login"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/doctor/:login": {
            "get": {
                "tags": [
                    "doctor"
                ],
                "summary": "Doctor found by login",
                "description": "Requires apikey",
                "operationId": "doctor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/loginString"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/doctor/specialization/:id": {
            "get": {
                "tags": [
                    "doctor"
                ],
                "summary": "Doctor and specialization found by doctor id",
                "description": "Requires apikey",
                "operationId": "doctorSpecialization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Id of doctor to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/doctor/signup": {
            "post": {
                "tags": [
                    "doctor"
                ],
                "summary": "Doctor register",
                "description": "Requires apikey",
                "operationId": "doctorSignup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/doctorSignup"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/doctor/datafillup": {
            "put": {
                "tags": [
                    "doctor"
                ],
                "summary": "Doctor profile data fill up",
                "description": "Requires apikey",
                "operationId": "doctorDataFillUp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/profileData"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/doctor/data/:id": {
            "get": {
                "tags": [
                    "doctor"
                ],
                "summary": "Doctor profile data",
                "description": "Requires apikey",
                "operationId": "doctorProfileData",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/profileData"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/doctor/login": {
            "post": {
                "tags": [
                    "doctor"
                ],
                "summary": "Doctor Login",
                "description": "Requires apikey",
                "operationId": "Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/login"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/loginDoctorResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "content": {}
                    }
                }
            }
        },
        "/doctor/:id": {
            "delete": {
                "tags": [
                    "doctor"
                ],
                "summary": "Delete doctor by worker's id",
                "description": "Requires apikey",
                "operationId": "doctorDelete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/login"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "patientSignup": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "haslo": {
                    "type": "string"
                },
                "imie": {
                    "type": "string"
                },
                "nazwisko": {
                    "type": "string"
                },
                "idplci": {
                    "type": "string"
                },
                "ulica": {
                    "type": "string"
                },
                "numerlokalu": {
                    "type": "string"
                },
                "miejscowosc": {
                    "type": "string"
                },
                "kodpocztowy": {
                    "type": "string"
                },
                "numertelefonu": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "loginString": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                }
            }
        },
        "doctorSignup": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "haslo": {
                    "type": "string"
                }
            }
        },
        "profileData": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "haslo": {
                    "type": "string"
                },
                "imie": {
                    "type": "string"
                },
                "nazwisko": {
                    "type": "string"
                },
                "idplci": {
                    "type": "string"
                },
                "ulica": {
                    "type": "string"
                },
                "numerlokalu": {
                    "type": "string"
                },
                "miejscowosc": {
                    "type": "string"
                },
                "kodpocztowy": {
                    "type": "string"
                },
                "numertelefonu": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "login": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "haslo": {
                    "type": "string"
                }
            }
        },
        "loginDoctorResponse": {
            "type": "object",
            "properties": {
                "doctorid": {
                    "type": "integer"
                },
                "workerid": {
                    "type": "integer"
                }
            }
        },
...
    }
}